Major fix to property set definition instantiation
The instantiation takes into account the parent object hierarchy better with, for
example, the info source being inherited correctly.
Minor fix to prevent null error for absent 'Order' property.

Also includes fix for deserialisation of propertyset/group/ies (and their
definitions) not repairing their parent relationships properly on
deserialisation.
This involves the addition of a custom serializer that runs the callbacks, which
it turns out the XmlSerializer does not do in contrast to the BinaryFormatter.

Also, major fix to XML value content class actually being able to have
content now, which some helpers to extract and set it.

Minor fixes to implicit/explicit casts on some classes.